Use super class for policy
