+++ python/unitytrainers/ppo/__init__.py
@@ -1,2 +1,3 @@
 from .models import *
 from .trainer import *
+from .policy import *
